# docs/API.yaml
openapi: 3.0.3
info:
  version: 3.0.1
  title: API
  description: "A list of the currently provided Interfaces to interact with the swarm, implementing file operations and sending messages"
security:
  - {}
...
servers:
  - url: "http://{apiRoot}:{port}/v1"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local hop node main API
      port:
        default: "1633"
        description: Service port provided in hop node config
  - url: "http://{apiRoot}:{port}"
    variables:
      apiRoot:
        default: "localhost"
        description: Base address of the local hop node main API
      port:
        default: "1633"
        description: Service port provided in hop node config
paths:
...
  "/hop":
    post:
      summary: "Upload file or a collection of files"
      description: "In order to upload a collection, user can send a multipart request with all the files populated in the form data with appropriate headers.\n\n
        User can also upload a tar file along with the swarm-collection header. This will upload the tar file after extracting the entire directory structure.\n\n
        If the swarm-collection header is absent, all requests (including tar files) are considered as single file uploads.\n\n
        A multipart request is treated as a collection regardless of whether the swarm-collection header is present. This means in order to serve single files
        uploaded as a multipart request, the swarm-index-document header should be used with the name of the file."
      tags:
        - HOP
      parameters:
        - in: query
          name: name
          schema:
            $ref: "Common.yaml#/components/schemas/FileName"
          required: false
          description: Filename when uploading single file
        - $ref: "Common.yaml#/components/parameters/SwarmTagParameter"
        - $ref: "Common.yaml#/components/parameters/SwarmPinParameter"
        - $ref: "Common.yaml#/components/parameters/SwarmEncryptParameter"
        - $ref: "Common.yaml#/components/parameters/ContentTypePreserved"
        - $ref: "Common.yaml#/components/parameters/SwarmCollection"
        - $ref: "Common.yaml#/components/parameters/SwarmIndexDocumentParameter"
        - $ref: "Common.yaml#/components/parameters/SwarmErrorDocumentParameter"
        - $ref: "Common.yaml#/components/parameters/SwarmPostageBatchId"
        - $ref: "Common.yaml#/components/parameters/SwarmDeferredUpload"
      requestBody:
        content:
          multipart/form-data:
            schema:
              properties:
                file:
                  type: array
                  items:
                    type: string
                    format: binary
          application/octet-stream:
            schema:
              type: string
              format: binary
          application/x-tar:
            schema:
              type: string
              format: binary
      responses:
        "201":
          description: Ok
          headers:
            "swarm-tag":
              $ref: "Common.yaml#/components/headers/SwarmTag"
            "etag":
              $ref: "Common.yaml#/components/headers/ETag"
          content:
            application/json:
              schema:
                $ref: "Common.yaml#/components/schemas/ReferenceResponse"
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "402":
          $ref: "Common.yaml#/components/responses/402"
        "403":
          $ref: "Common.yaml#/components/responses/GatewayForbidden"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
  "/hop/{reference}":
    patch:
      summary: "Reupload a root hash to the network; deprecated: use /stewardship/{reference} instead"
      deprecated: true
      tags:
        - HOP
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "Common.yaml#/components/schemas/SwarmReference"
          required: true
          description: "Root hash of content (can be of any type: collection, file, chunk)"
      responses:
        "200":
          description: Ok
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
    get:
      summary: "Get file or index document from a collection of files"
      tags:
        - HOP
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "Common.yaml#/components/schemas/SwarmReference"
          required: true
          description: Swarm address of content
      responses:
        "200":
          description: Ok
          # "swarm-feed-index":
          #   $ref: "Common.yaml#/components/headers/SwarmFeedIndex"
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          $ref: "Common.yaml#/components/responses/400"
        "404":
          $ref: "Common.yaml#/components/responses/404"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
  "/hop/{reference}/{path}":
    get:
      summary: "Get referenced file from a collection of files"
      tags:
        - HOP
      parameters:
        - in: path
          name: reference
          schema:
            $ref: "Common.yaml#/components/schemas/SwarmReference"
          required: true
          description: Swarm address of content
        - in: path
          name: path
          schema:
            type: string
          required: true
          description: Path to the file in the collection.
      responses:
        "200":
          description: Ok
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

        "400":
          $ref: "Common.yaml#/components/responses/400"
        "404":
          $ref: "Common.yaml#/components/responses/404"
        "500":
          $ref: "Common.yaml#/components/responses/500"
        default:
          description: Default response
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SecurityToken